fix some of magd's feedback
